refactor: refactor project table
